requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# LLM & MCP
openai==1.3.5
mcp==0.1.0

# Image processing (for visual tag extraction)
Pillow==10.1.0
torch==2.1.0
transformers==4.35.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.0

